split module in two
