+++ auto-pos-creator/src/external_sc_interactions/mod.rs
@@ -1,4 +1,3 @@
 pub mod farm_actions;
 pub mod metastaking_actions;
-pub mod multi_contract_interactions;
 pub mod pair_actions;
